% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_puzzle}
\alias{demo_puzzle}
\alias{hi_puzzle}
\title{Hi puzzle}
\format{
a puzzle S7 object

A simple sample puzzle to show how the package works.
Hint: it's a glider from Conway's Game of Life.

a puzzle S7 object

A simple  asymmetric sample puzzle to show how the package works.
Hint: it's block text Hi! in a minimal piizel font
}
\usage{
demo_puzzle

hi_puzzle
}
\description{
Hi puzzle
}
\keyword{datasets}
