% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hi_puzzle}
\alias{hi_puzzle}
\title{Hi puzzle}
\format{
a puzzle S7 object. See \code{puzzle} for details.
}
\usage{
hi_puzzle
}
\description{
Hi puzzle
}
\details{
A simple  asymmetric sample puzzle to show how the package works.
Hint: it's block text Hi! in a minimal piizel font
}
\keyword{datasets}
