% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonogramr-package.R
\docType{package}
\name{nonogramr-package}
\alias{nonogramr}
\alias{nonogramr-package}
\title{nonogramr: Tools for nonograms}
\description{
Tools for solving nonogram puzzles in R.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{nonogramr@AlexAxthelm.com} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

}
\keyword{internal}
